put new test in its own file
