+++ src/serac/physics/tests/CMakeLists.txt
@@ -28,6 +28,7 @@ set(solver_tests
     solid.cpp
     solid_periodic.cpp
     solid_shape.cpp
+    solid_robin_condition.cpp
     thermal_shape.cpp
     thermal_mechanics.cpp
     )
